feat(timestamped-comments): addressing PR comments
