+++ src/elements/content-sidebar/activity-feed/common/activity-message/messages.js
@@ -27,7 +27,7 @@ const messages = defineMessages({
         defaultMessage: '\\ (edited)',
         description: 'Label indicating that message was edited, should be lowercase, should have escaped leading space',
     },
-    commentMessageTimestampLabel: {
+    activityMessageTimestampLabel: {
         id: 'be.contentSidebar.activityFeed.common.timestampLabel',
         defaultMessage: 'Seek to video timestamp',
         description: 'Accessibility label for timestamp button that allows jumping to a specific time in media',
