+++ src/styles/constants/_colors.scss
@@ -7,7 +7,6 @@ $countBadgeBackground: #e9163c !default;
 
 // Box/Primary Blues
 $bdl-box-blue: #0061d5 !default;
-$bdl-box-blue-110: #0057c0 !default;
 $bdl-box-blue-80: #3381dd !default;
 $bdl-box-blue-50: #7fb0ea !default;
 $bdl-box-blue-40: #99c0ee !default;
