+++ src/styles/variables.json
@@ -3,7 +3,6 @@
     "black": "#000",
     "countBadgeBackground": "#e9163c",
     "bdl-box-blue": "#0061d5",
-    "bdl-box-blue-110": "#0057C0",
     "bdl-box-blue-80": "#3381dd",
     "bdl-box-blue-50": "#7fb0ea",
     "bdl-box-blue-40": "#99c0ee",
