+++ src/utils/timestamp.js.flow
@@ -4,4 +4,3 @@ import type { IntlShape } from 'react-intl';
 declare export function convertTimestampToSeconds(timestamp: number): number;
 declare export function convertMillisecondsToHMMSS(timestampInMilliseconds: number): string;
 declare export function convertSecondsToHMMSS(seconds: number): string;
-declare export function formatTimestamp(text: string, timestamp: string, intl: IntlShape): React$Element<any> | string;
